Extract Statistics::General from GeneralController
Reduce complexity of controller (and its tests) by extracting logic to a
more appropriate model.

Allows us to remove a non-controller method from ApplicationController.

Also allows easier access from the command line:

    bin/rails runner "pp Statistics::General.new.to_h"
